Create base class/factory for Solr + ILS commands.
